#36174 Move template resolver from block to model
